chain.*: file the device count as int, and drop a stray escape in a message
QDomElement::setAttribute with a container size() is ambiguous between the int and
double overloads under Qt6 (EffectChain casts for the same reason), and a
mechanism string carried a needless backslash before an apostrophe.

+++ src/core/ControlChainPresetSupport.cpp
@@ -164,7 +164,7 @@ QByteArray controlChainPresetDocument(const QString& name,
 	QDomElement root = doc.createElement(chainPresetRootName());
 	root.setAttribute(QStringLiteral("version"), 1);
 	root.setAttribute(QStringLiteral("name"), name);
-	root.setAttribute(QStringLiteral("devices"), devices.size());
+	root.setAttribute(QStringLiteral("devices"), static_cast<int>(devices.size()));
 	doc.appendChild(root);
 
 	for (const ControlChainPresetDevice& device : devices)
